Move snprintf and arraysize definitions to header file
